Revert "Set BiSet store to ulong"
This reverts commit 5f4549ddaa77232ee3586ae4ac41b40969fcab14.

+++ aoc/BitSet.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Numerics;
 using System.Runtime.CompilerServices;
-using Store = System.UInt64;
+using Store = System.Int64;
 
 namespace aoc
 {
@@ -108,7 +108,7 @@ namespace aoc
         {
             int count = 0;
             for (int i = 0; i < _bits.Length; i++)
-                count += BitOperations.PopCount(_bits[i]);
+                count += BitOperations.PopCount((ulong)_bits[i]);
             return count;
         }
 
@@ -159,7 +159,7 @@ namespace aoc
         {
             for (int i = _bits.Length - 1; i >= 0; i--)
                 if ((store = _bits[i]) != 0)
-                    return i << Shift | BitOperations.Log2(store);
+                    return i << Shift | BitOperations.Log2((ulong)store);
             store = 0;
             return -1;
         }
@@ -168,10 +168,10 @@ namespace aoc
         public readonly int LastSet(int index, out Store store)
         {
             if ((store = _bits[index >> Shift] & ((One << index) - 1)) != 0)
-                return index & ~Mask | BitOperations.Log2(store);
+                return index & ~Mask | BitOperations.Log2((ulong)store);
             for (int i = (index >> Shift) - 1; i >= 0; i--)
                 if ((store = _bits[i]) != 0)
-                    return i << Shift | BitOperations.Log2(store);
+                    return i << Shift | BitOperations.Log2((ulong)store);
             return -1;
         }
 
@@ -204,10 +204,10 @@ namespace aoc
         public readonly int PreviousSet(int index, ref Store store)
         {
             if ((store &= (One << index) - 1) != 0)
-                return index & ~Mask | BitOperations.Log2(store);
+                return index & ~Mask | BitOperations.Log2((ulong)store);
             for (int i = (index >> Shift) - 1; i >= 0; i--)
                 if ((store = _bits[i]) != 0)
-                    return i << Shift | BitOperations.Log2(store);
+                    return i << Shift | BitOperations.Log2((ulong)store);
             return -1;
         }
